Create Window Privilege Escalation via Automated Script
We all know that after compromising the victim's machine we have a low-privileges shell that we want to escalate into a higher-privileged shell and this process is 
known as Privilege Escalation. There are some scripts for Windows that may come in useful when trying to escalate privileges on a target system. This is generally 
aimed at enumeration rather than specific vulnerabilities/exploits. This type of script could save your much time.